finished the array logic

+++ input/input_main.h
@@ -74,7 +74,7 @@ along with Czech_mate.  If not, see <http://www.gnu.org/licenses/>.
 	};
 	class input_t{
 		private:
-			void input_find_key(unsigned int*, const input_buffer_t*);
+			void input_find_key(unsigned long int*, const input_buffer_t*);
 			void input_update_key(input_buffer_t*);
 			void input_parse_key_up(SDL_Event);
 			void input_parse_key_down(SDL_Event);
